requirements.txt
django-model-utils==4.0.0
Django==3.0.6
djangorestframework==3.11.0
psycopg2-binary==2.8.5
pytz==2020.1
pytest-django==4.1.0
requests==2.25.1
